Fix screen sharing in Sway.
- Remove Gnome (I no longer need it)
- Disable Hyprland (screensharing broke and PyCharm isn't working as
  expected)
- Remove `xdg-desktop-portal-hyprland` to fix screen sharing in Sway

In order to use Hyprland from now on you need to enable the xdg package.
I might come back to it at a later point.

What was the issue with PyCharm?
Floating windows were popping up anywhere  on the screen. Sometimes you
could only see the corner of the window and you had to drag it to the
center of the screen.
Some floating windows were just 50% opacity and blurry and did not hold
any content.