Fix space mapping in neo-tree.
After the last update, pressing the `<space>` key in neo-tree was no
longer opening the command prompt from AstroNvim, but toggled the file
tree node (after a delay).
The delay was there, because neo-tree was waiting for another function
to react first, which it didn't.
I figured out, that disabling the key completely will resolve this
completely also for future updates.