Remove AMD GPU and Prime.
Initially I wanted to use the AMD GPU for Wayland and Nvidia for
Gaming / LLM only.
However, Sway or Wayland did not like that very much and everything was
still running on the AMD GPU to ~50%.
I did not find a fix, but figured out since Sway is running on the
Nvidia GPU, it should run fine on Nvidia only too and it does (back then
it didn't).